Overflowing a short length so it goes negative and screws up the byte…

+++ Source/com/drew/metadata/photoshop/PhotoshopReader.java
@@ -110,7 +110,7 @@ public class PhotoshopReader implements JpegSegmentMetadataReader
                 pos += 2;
 
                 // A variable number of bytes holding a pascal string (two leading bytes for length).
-                short descriptionLength = reader.getUInt8();
+                int descriptionLength = reader.getUInt8();
                 pos += 1;
                 // Some basic bounds checking
                 if (descriptionLength < 0 || descriptionLength + pos > length)
